Move all services to services/ and remove "svc-" prefix
