+++ .circleci/config.yml
@@ -8,8 +8,8 @@ jobs:
       - checkout
       - run: go get -u github.com/golang/dep/cmd/dep
       - run: cd gateway && dep ensure -vendor-only
-      - run: cd svc-add && dep ensure -vendor-only
-      - run: cd svc-string && dep ensure -vendor-only
+      - run: cd services/add && dep ensure -vendor-only
+      - run: cd services/string && dep ensure -vendor-only
       - run: go test $(go list ./... | grep -v /vendor/)
   build-and-deploy:
     working_directory: /go/src/github.com/lukyth/try-go-kit/
@@ -46,8 +46,8 @@ jobs:
           name: Deploy Master to GKE
           command: |
             ./deploy.sh gateway
-            ./deploy.sh svc-add
-            ./deploy.sh svc-string
+            ./deploy.sh services/add
+            ./deploy.sh services/string
 workflows:
   version: 2
   build_accept_deploy:
