+++ deploy.sh
@@ -4,7 +4,7 @@
 set -e
 
 SERVICE_NAME=$1
-IMAGE_NAME=gcr.io/firsttry-178217/try-go-kit/$SERVICE_NAME
+IMAGE_NAME=gcr.io/firsttry-178217/try-go-kit/services/$SERVICE_NAME
 IMAGE_NAME_WITH_TAG=$IMAGE_NAME:$CIRCLE_SHA1
 IMAGE_NAME_WITH_LATEST_TAG=$IMAGE_NAME:latest
