+++ services/add/.gitignore
@@ -1,0 +1,2 @@
+vendor/
+cmd/add/add
