+++ services/message/README.md
@@ -1,0 +1,19 @@
+# Message Service
+
+## Installation
+### Install dependencies
+```bash
+dep ensure
+```
+### Build a service binary
+```bash
+./build.sh
+```
+
+## Usage
+```bash
+./bin/message
+```
+```bash
+curl localhost:8080
+```
