+++ services/message/build.sh
@@ -1,0 +1,3 @@
+#!/bin/bash
+
+go build -o bin/message cmd/message.go
