+++ services/string/.gitignore
@@ -1,0 +1,2 @@
+vendor/
+cmd/string/string
