+++ services/string/pkg/service/instrumenting.go
