--- svc-add/.gitignore
@@ -1,2 +1,0 @@
-vendor/
-cmd/svc-add/svc-add
