--- svc-add/cmd/svc-add/svc-add.go
@@ -1,124 +1,0 @@
-package main
-
-import (
-	"flag"
-	"fmt"
-	"net"
-	"net/http"
-	"os"
-	"os/signal"
-	"syscall"
-	"text/tabwriter"
-
-	"github.com/oklog/oklog/pkg/group"
-	"github.com/prometheus/client_golang/prometheus/promhttp"
-
-	"github.com/go-kit/kit/log"
-
-	"github.com/lukyth/try-go-kit/svc-add/pkg/endpoint"
-	"github.com/lukyth/try-go-kit/svc-add/pkg/service"
-	"github.com/lukyth/try-go-kit/svc-add/pkg/transport"
-)
-
-func main() {
-	// Define our flags.
-	fs := flag.NewFlagSet("svc-add", flag.ExitOnError)
-	var (
-		httpAddr  = fs.String("http-addr", ":8080", "HTTP listen address")
-		debugAddr = fs.String("debug.addr", ":8081", "Debug and metrics listen address")
-	)
-	fs.Usage = usageFor(fs, os.Args[0]+" [flags]")
-	fs.Parse(os.Args[1:])
-
-	// Create a single logger, which we'll use and give to other components.
-	var logger log.Logger
-	{
-		logger = log.NewLogfmtLogger(os.Stderr)
-		logger = log.With(logger, "ts", log.DefaultTimestampUTC)
-		logger = log.With(logger, "caller", log.DefaultCaller)
-	}
-
-	http.DefaultServeMux.Handle("/metrics", promhttp.Handler())
-
-	var (
-		service     = service.New(logger)
-		endpoints   = endpoint.New(service, logger)
-		httpHandler = transport.NewHTTPHandler(endpoints, logger)
-	)
-
-	// Now we're to the part of the func main where we want to start actually
-	// running things, like servers bound to listeners to receive connections.
-	//
-	// The method is the same for each component: add a new actor to the group
-	// struct, which is a combination of 2 anonymous functions: the first
-	// function actually runs the component, and the second function should
-	// interrupt the first function and cause it to return. It's in these
-	// functions that we actually bin the Go kit server/handler structs to the
-	// concrete transports and start them running.
-	//
-	// Putting each component into its own block is mostly for aesthetics: it
-	// clearly demarcates the scope in which each listener/socket may be used.
-	var g group.Group
-	{
-		// The debug listener mounts the http.DefaultServeMux, and serves up
-		// stuff like the Prometheus metrics route, the Go debug and profiling
-		// routes, and so on.
-		debugListener, err := net.Listen("tcp", *debugAddr)
-		if err != nil {
-			logger.Log("transport", "debug/HTTP", "during", "Listen", "err", err)
-			os.Exit(1)
-		}
-		g.Add(func() error {
-			logger.Log("transport", "debug/HTTP", "addr", *debugAddr)
-			return http.Serve(debugListener, http.DefaultServeMux)
-		}, func(error) {
-			debugListener.Close()
-		})
-	}
-	{
-		// The HTTP listener mounts the Go kit HTTP handler we created.
-		httpListener, err := net.Listen("tcp", *httpAddr)
-		if err != nil {
-			logger.Log("transport", "HTTP", "during", "Listen", "err", err)
-			os.Exit(1)
-		}
-		g.Add(func() error {
-			logger.Log("transport", "HTTP", "addr", *httpAddr)
-			return http.Serve(httpListener, httpHandler)
-		}, func(error) {
-			httpListener.Close()
-		})
-	}
-	{
-		// This function just sits and waits for ctrl-C.
-		cancelInterrupt := make(chan struct{})
-		g.Add(func() error {
-			c := make(chan os.Signal, 1)
-			signal.Notify(c, syscall.SIGINT, syscall.SIGTERM)
-			select {
-			case sig := <-c:
-				return fmt.Errorf("received signal %s", sig)
-			case <-cancelInterrupt:
-				return nil
-			}
-		}, func(error) {
-			close(cancelInterrupt)
-		})
-	}
-	logger.Log("exit", g.Run())
-}
-
-func usageFor(fs *flag.FlagSet, short string) func() {
-	return func() {
-		fmt.Fprintf(os.Stderr, "USAGE\n")
-		fmt.Fprintf(os.Stderr, "  %s\n", short)
-		fmt.Fprintf(os.Stderr, "\n")
-		fmt.Fprintf(os.Stderr, "FLAGS\n")
-		w := tabwriter.NewWriter(os.Stderr, 0, 2, 2, ' ', 0)
-		fs.VisitAll(func(f *flag.Flag) {
-			fmt.Fprintf(w, "\t-%s %s\t%s\n", f.Name, f.DefValue, f.Usage)
-		})
-		w.Flush()
-		fmt.Fprintf(os.Stderr, "\n")
-	}
-}
