--- svc-add/pkg/endpoint/middleware.go
@@ -1,59 +1,0 @@
-package endpoint
-
-import (
-	"context"
-	"fmt"
-	"time"
-
-	stdprometheus "github.com/prometheus/client_golang/prometheus"
-
-	"github.com/go-kit/kit/endpoint"
-	"github.com/go-kit/kit/log"
-	"github.com/go-kit/kit/metrics"
-	"github.com/go-kit/kit/metrics/prometheus"
-)
-
-// GenerateInstrumentingMetric returns an endpoint-level metrics.
-func GenerateInstrumentingMetric() metrics.Histogram {
-	// Endpoint-level metrics.
-	duration := prometheus.NewSummaryFrom(stdprometheus.SummaryOpts{
-		Namespace: "try_go_kit",
-		Subsystem: "svc_add",
-		Name:      "request_duration_seconds",
-		Help:      "Request duration in seconds.",
-	}, []string{"method", "success"})
-
-	return duration
-}
-
-// InstrumentingMiddleware returns an endpoint middleware that records
-// the duration of each invocation to the passed histogram. The middleware adds
-// a single field: "success", which is "true" if no error is returned, and
-// "false" otherwise.
-func InstrumentingMiddleware(duration metrics.Histogram) endpoint.Middleware {
-	return func(next endpoint.Endpoint) endpoint.Endpoint {
-		return func(ctx context.Context, request interface{}) (response interface{}, err error) {
-
-			defer func(begin time.Time) {
-				duration.With("success", fmt.Sprint(err == nil)).Observe(time.Since(begin).Seconds())
-			}(time.Now())
-			return next(ctx, request)
-
-		}
-	}
-}
-
-// LoggingMiddleware returns an endpoint middleware that logs the
-// duration of each invocation, and the resulting error, if any.
-func LoggingMiddleware(logger log.Logger) endpoint.Middleware {
-	return func(next endpoint.Endpoint) endpoint.Endpoint {
-		return func(ctx context.Context, request interface{}) (response interface{}, err error) {
-
-			defer func(begin time.Time) {
-				logger.Log("transport_error", err, "took", time.Since(begin))
-			}(time.Now())
-			return next(ctx, request)
-
-		}
-	}
-}
