--- svc-add/pkg/transport/http.go
@@ -1,74 +1,0 @@
-package transport
-
-import (
-	"context"
-	"encoding/json"
-	"net/http"
-
-	"github.com/go-kit/kit/log"
-	httptransport "github.com/go-kit/kit/transport/http"
-
-	addendpoint "github.com/lukyth/try-go-kit/svc-add/pkg/endpoint"
-	"github.com/lukyth/try-go-kit/svc-add/pkg/service"
-)
-
-// NewHTTPHandler returns an HTTP handler that makes a set of endpoints
-// available on predefined paths.
-func NewHTTPHandler(endpoints addendpoint.Set, logger log.Logger) http.Handler {
-	m := http.NewServeMux()
-	m.Handle("/sum", httptransport.NewServer(
-		endpoints.SumEndpoint,
-		decodeHTTPSumRequest,
-		encodeHTTPGenericResponse,
-	))
-	m.Handle("/concat", httptransport.NewServer(
-		endpoints.ConcatEndpoint,
-		decodeHTTPConcatRequest,
-		encodeHTTPGenericResponse,
-	))
-	return m
-}
-
-func errorEncoder(_ context.Context, err error, w http.ResponseWriter) {
-	w.WriteHeader(err2code(err))
-	json.NewEncoder(w).Encode(errorWrapper{Error: err.Error()})
-}
-
-func err2code(err error) int {
-	switch err {
-	case service.ErrTwoZeroes, service.ErrMaxSizeExceeded, service.ErrIntOverflow:
-		return http.StatusBadRequest
-	}
-	return http.StatusInternalServerError
-}
-
-type errorWrapper struct {
-	Error string `json:"error"`
-}
-
-// decodeHTTPSumRequest is a transport/http.DecodeRequestFunc that decodes a
-// JSON-encoded sum request from the HTTP request body.
-func decodeHTTPSumRequest(_ context.Context, r *http.Request) (interface{}, error) {
-	var req addendpoint.SumRequest
-	err := json.NewDecoder(r.Body).Decode(&req)
-	return req, err
-}
-
-// decodeHTTPConcatRequest is a transport/http.DecodeRequestFunc that decodes a
-// JSON-encoded concat request from the HTTP request body.
-func decodeHTTPConcatRequest(_ context.Context, r *http.Request) (interface{}, error) {
-	var req addendpoint.ConcatRequest
-	err := json.NewDecoder(r.Body).Decode(&req)
-	return req, err
-}
-
-// encodeHTTPGenericResponse is a transport/http.EncodeResponseFunc that encodes
-// the response as JSON to the response writer.
-func encodeHTTPGenericResponse(ctx context.Context, w http.ResponseWriter, response interface{}) error {
-	if f, ok := response.(addendpoint.Failer); ok && f.Failed() != nil {
-		errorEncoder(ctx, f.Failed(), w)
-		return nil
-	}
-	w.Header().Set("Content-Type", "application/json; charset=utf-8")
-	return json.NewEncoder(w).Encode(response)
-}
