--- svc-message/README.md
@@ -1,19 +1,0 @@
-# Message Service
-
-## Installation
-### Install dependencies
-```bash
-dep ensure
-```
-### Build a service binary
-```bash
-./build.sh
-```
-
-## Usage
-```bash
-./bin/svc-message
-```
-```bash
-curl localhost:8080
-```
