--- svc-message/build.sh
@@ -1,3 +1,0 @@
-#!/bin/bash
-
-go build -o bin/svc-message cmd/svc-message.go
