--- svc-string/.gitignore
@@ -1,2 +1,0 @@
-vendor/
-cmd/svc-string/svc-string
